requirements.txt
numpy==2.1.3
pillow==11.1.0
scipy==1.15.2
tqdm==4.67.1